pyproject.toml
[project]
name = "resume-ai"
version = "0.1.0"
description = ""
authors = [{ name = "Austin Lynn" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi[standard] (>=0.115.6,<0.116.0)", "pydantic-settings (>=2.7.1,<3.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.4"
pytest-mypy-plugins = "^3.2.0"
mypy = "^1.14.1"
pyright = "^1.1.391"
black = "^24.10.0"
isort = "^5.13.2"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["app"]

[tool.isort]
profile = "black"
